% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossurr-package.R
\docType{package}
\name{crossurr-package}
\alias{crossurr}
\alias{crossurr-package}
\title{crossurr}
\description{
The main functions of this package are \code{xf_surrogate} and \code{xfr_surrogate}
}
\author{
\strong{Maintainer}: Denis Agniel \email{dagniel@rand.org}

Authors:
\itemize{
  \item Boris P. Hejblum \email{boris.hejblum@u-bordeaux.fr}
  \item Layla Parast \email{parast@austin.utexas.edu}
}

}
\keyword{internal}
