% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lasso}
\alias{lasso}
\title{lasso}
\usage{
lasso(
  x = NULL,
  y = NULL,
  data = NULL,
  newX = NULL,
  newX0 = NULL,
  newX1 = NULL,
  relax = TRUE,
  ps_fit = FALSE,
  ...
)
}
\description{
lasso
}
\keyword{internal}
