% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric-models.R
\name{ols}
\alias{ols}
\title{Ordinary Least Squares}
\usage{
ols(
  x = NULL,
  y = NULL,
  data = NULL,
  test_data = NULL,
  test_data0 = NULL,
  test_data1 = NULL,
  ...
)
}
\description{
Ordinary Least Squares
}
\keyword{internal}
