% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{A simple function to simulate example data.}
\usage{
sim_data(n, p)
}
\arguments{
\item{n}{number of simulated observations}

\item{p}{number of simulated variables}
}
\description{
A simple function to simulate example data.
}
