% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfit_dr.R
\name{xfit_dr}
\alias{xfit_dr}
\title{Title}
\usage{
xfit_dr(
  ds,
  x,
  y,
  a,
  K = 5,
  outcome_learners = NULL,
  ps_learners = outcome_learners,
  interaction_model = TRUE,
  trim_at = 0.05,
  outcome_family = gaussian(),
  mthd = "superlearner",
  ncores = parallel::detectCores() - 1,
  ...
)
}
\description{
Title
}
\keyword{internal}
