% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{A simple function to simulate example data.}
\usage{
sim_data(n, p)
}
\arguments{
\item{n}{number of simulated observations}

\item{p}{number of simulated variables}
}
\value{
toy dataset used for demonstrating the methods with outcome \code{y}, treatment \code{a}, covariates \code{x.1, x.2}, and surrogates \code{s.1, s.2, ...}
}
\description{
A simple function to simulate example data.
}
