% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xf_surrogate.R
\name{xf_surrogate}
\alias{xf_surrogate}
\title{A function for ...}
\usage{
xf_surrogate(
  ds,
  x = NULL,
  s,
  y,
  a,
  K = 5,
  outcome_learners = NULL,
  ps_learners = outcome_learners,
  interaction_model = TRUE,
  trim_at = 0.05,
  outcome_family = gaussian(),
  mthd = "superlearner",
  n_ptb = 0,
  ncores = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{ds}{a \code{data.frame}.}

\item{x}{names any other covariates (eg. age, sex, etc). Default is \code{NULL}.}

\item{s}{names surrogates}

\item{y}{names outcome}

\item{a}{treatment variable name (eg. groups). Expect a binary variable made of \code{1}s and \code{0}s.}

\item{K}{number of folds for cross validation. Default is \code{5}}

\item{outcome_learners}{Default is \code{NULL}}

\item{ps_learners}{only used for the superlearner}

\item{interaction_model}{logical indicating wether an interaction model should be considered.
Default is \code{TRUE}}

\item{trim_at}{Default is \code{0.05}}

\item{outcome_family}{Default is \code{'gaussian'}. Other choice is \code{'binomial'} for binary outcome.}

\item{mthd}{Regression method. Default is \code{'superlearner'}, other choice
is \code{'lasso'}.}

\item{n_ptb}{Number of perturbations. Default is \code{0} which means asymptotics}

\item{ncores}{number of cpus used for parallel computations. Default is \code{parallel::detectCores()-1}}

\item{...}{additional parameters (in particular for super_learner)}
}
\description{
A function for ...
}
